+++ tests/test_converter.py
@@ -239,7 +239,7 @@ class TestConverter:
 
         with pytest.raises(ValueError, match='PennyLane does not support expressions'):
             with recorder:
-                quantum_circuit(params={theta: qml.variable.Variable(0), phi: qml.variable.Variable(1)})
+                quantum_circuit(params={theta: 0, phi: 1})
 
     def test_extra_parameters_were_passed(self, recorder):
         """Tests that loading raises an error when extra parameters were
@@ -247,6 +247,8 @@ class TestConverter:
 
         theta = Parameter('θ')
         phi = Parameter('φ')
+        x = np.tensor(0.5, requires_grad=False)
+        y = np.tensor(0.3, requires_grad=False)
 
         qc = QuantumCircuit(3, 1)
 
@@ -254,7 +256,7 @@ class TestConverter:
 
         with pytest.raises(QiskitError):
             with recorder:
-                quantum_circuit(params={theta: 0.5, phi: 0.3})
+                quantum_circuit(params={theta: x, phi: y})
 
     @pytest.mark.parametrize("qiskit_operation, pennylane_name", [(QuantumCircuit.crx, "CRX"), (QuantumCircuit.crz, "CRZ"), (QuantumCircuit.cry, "CRY")])
     def test_controlled_rotations(self, qiskit_operation, pennylane_name, recorder):
@@ -588,8 +590,8 @@ class TestConverter:
         """Tests the load method for a QuantumCircuit raises a QiskitError,
         if the wrong type of arguments were passed."""
 
-        theta = Parameter('θ')
-        angle = 'some_string_instead_of_an_angle'
+        theta = Parameter("θ")
+        angle = np.tensor("some_string_instead_of_an_angle", requires_grad=False)
 
         qc = QuantumCircuit(3, 1)
         qc.rz(theta, [0])
@@ -605,7 +607,7 @@ class TestConverter:
         that are not required were passed."""
 
         theta = Parameter('θ')
-        angle = 0.5
+        angle = np.tensor(0.5, requires_grad=False)
 
         qc = QuantumCircuit(3, 1)
         qc.z([0])